DCS-1247 Refactoring large module
