+++ server/services/conditionsService.ts
@@ -1,7 +1,8 @@
 import moment from 'moment'
 import { formatConditionsInput, getConditionText, formatConditionsText } from './utils/conditionsFormatter'
-import { getIn, isEmpty, interleave } from '../utils/functionalHelpers'
-import { getAdditionalConditionsConfig, standardConditions, multiFields } from './config/conditionsConfig'
+import { isEmpty } from '../utils/functionalHelpers'
+import * as builder from './licenceWithConditionsBuilder'
+import { getAdditionalConditionsConfig, standardConditions } from './config/conditionsConfig'
 import { AdditionalConditions, Licence } from '../data/licenceTypes'
 
 export default function createConditionsService() {
@@ -15,7 +16,7 @@ export default function createConditionsService() {
       return { standardConditions: standardConditionsText, additionalConditions: [] }
     }
 
-    const conditions = populateLicenceWithApprovedConditions(licence).licenceConditions
+    const conditions = builder.populateLicenceWithApprovedConditions(licence).licenceConditions
 
     // conditions can be either the original licenceConditions object or an array of objects if
     // populateLicenceWithApprovedConditions did its thing.
@@ -60,41 +61,6 @@ export default function createConditionsService() {
     return formatConditionsInput(requestBody, selectedConditionsConfig)
   }
 
-  function populateLicenceWithApprovedConditions(licence: Licence) {
-    return populateLicenceWithConditions(licence, null, true)
-  }
-
-  function populateLicenceWithConditions(licence: Licence, errors = {}, approvedOnly = false) {
-    // could be undefined, 'No' or 'Yes'
-    if (licence?.licenceConditions?.standard?.additionalConditionsRequired !== 'Yes') {
-      return licence
-    }
-
-    const licenceAdditionalConditions = licence?.licenceConditions?.additional
-    const bespokeConditions = licence?.licenceConditions?.bespoke || []
-    const conditionsOnLicence = !isEmpty(licenceAdditionalConditions) || bespokeConditions.length > 0
-
-    if (!conditionsOnLicence) {
-      return licence
-    }
-
-    const conditionIdsSelected = Object.keys(licenceAdditionalConditions)
-    const selectedConditionsConfig = getAdditionalConditionsConfig('V1').filter((condition) =>
-      conditionIdsSelected.includes(condition.id)
-    )
-
-    const abuseAndBehavioursConditions = licence?.licenceConditions?.additional?.COMPLYREQUIREMENTS?.abuseAndBehaviours
-
-    if (Array.isArray(abuseAndBehavioursConditions)) {
-      Object.assign(
-        abuseAndBehavioursConditions,
-        abuseAndBehavioursConditions.map((condition, index) => (index > 0 ? ` ${condition}` : condition))
-      )
-    }
-
-    return populateAdditionalConditionsAsObject(licence, selectedConditionsConfig, errors, approvedOnly)
-  }
-
   function createConditionsObjectForLicence(additional, bespoke) {
     if (isEmpty(additional)) {
       return { additional: {}, bespoke }
@@ -123,72 +89,8 @@ export default function createConditionsService() {
     }, {})
   }
 
-  function populateAdditionalConditionsAsObject(
-    rawLicence: Licence,
-    selectedConditionsConfig,
-    inputErrors = {},
-    approvedOnly = false
-  ) {
-    const { additional, bespoke } = rawLicence.licenceConditions
-
-    const additionalConditionsJustification =
-      rawLicence?.licenceConditions?.conditionsSummary?.additionalConditionsJustification
-    const getObjectForAdditional = createAdditionalMethod(rawLicence, selectedConditionsConfig, inputErrors)
-
-    const populatedAdditional = Object.keys(additional)
-      .sort(orderForView(getAdditionalConditionsConfig('V1').map((condition) => condition.id)))
-      .map(getObjectForAdditional)
-
-    const populatedBespoke = bespoke ? bespoke.map(getObjectForBespoke) : []
-    const selectedBespoke = approvedOnly
-      ? populatedBespoke.filter((condition) => condition.approved === 'Yes')
-      : populatedBespoke
-
-    return {
-      ...rawLicence,
-      licenceConditions: [...populatedAdditional, ...selectedBespoke],
-      additionalConditionsJustification,
-    }
-  }
-
-  function createAdditionalMethod(rawLicence: Licence, selectedConditions, inputErrors) {
-    return (condition) => {
-      const selectedCondition = selectedConditions.find((selected) => String(selected.id) === String(condition))
-      const userInput = rawLicence?.licenceConditions?.additional?.[condition]
-      const userErrors = inputErrors?.[condition]
-      const content = getContentForCondition(selectedCondition, userInput, userErrors)
-
-      return {
-        content,
-        group: selectedCondition.group_name,
-        subgroup: selectedCondition.subgroup_name,
-        id: selectedCondition.id,
-        inputRequired: !!selectedCondition.user_input,
-        approved: undefined,
-      }
-    }
-  }
-
-  function getContentForCondition(selectedCondition, userInput, userErrors) {
-    const userInputName = selectedCondition.user_input
-
-    return userInputName
-      ? injectUserInputAsObject(selectedCondition, userInput, userErrors)
-      : [{ text: selectedCondition.text }]
-  }
-
-  function getObjectForBespoke(condition, index) {
-    return {
-      content: [{ text: condition.text }],
-      group: 'Bespoke',
-      subgroup: null,
-      id: `bespoke-${index}`,
-      approved: condition.approved,
-    }
-  }
-
   function getNonStandardConditions(licence) {
-    let { licenceConditions = [] } = populateLicenceWithConditions(licence)
+    let { licenceConditions = [] } = builder.populateLicenceWithConditions(licence)
     if (!Array.isArray(licenceConditions)) {
       licenceConditions = []
     }
@@ -213,64 +115,6 @@ export default function createConditionsService() {
     }
   }
 
-  function injectUserInputAsObject(condition, userInput, userErrors) {
-    const conditionName = condition.user_input
-    const conditionText = condition.text
-    const fieldPositionObject = condition.field_position
-    const { userContent, errors } = condition.manipulateInput
-      ? condition.manipulateInput(userInput, userErrors)
-      : { userContent: userInput, errors: userErrors }
-
-    if (multiFields[conditionName]) {
-      return injectMultiFieldsAsObject(userContent, conditionText, errors, multiFields[conditionName])
-    }
-
-    return injectUserInputStandardAsObject(userContent, conditionText, fieldPositionObject, errors)
-  }
-
-  function injectUserInputStandardAsObject(userInput, conditionText, fieldPositionObject, userErrors) {
-    const fieldNames = Object.keys(fieldPositionObject)
-    const splitConditionText = conditionText.split(betweenBrackets).filter((text) => text)
-    const reducer = injectVariablesIntoViewObject(fieldNames, fieldPositionObject, userInput, userErrors)
-    return splitConditionText.reduce(reducer, [])
-  }
-
-  function injectVariablesIntoViewObject(fieldNames, fieldPositionObject, userInput, userErrors) {
-    return (conditionArray, textSegment, index) => {
-      const fieldNameForPlaceholder = fieldNames.find((field) => String(fieldPositionObject[field]) === String(index))
-      if (!fieldNameForPlaceholder) {
-        return [...conditionArray, { text: textSegment }]
-      }
-
-      const inputError = getIn(userErrors, [fieldNameForPlaceholder])
-      if (inputError) {
-        return [...conditionArray, { text: textSegment }, { error: `[${inputError}]` }]
-      }
-
-      const inputtedData = getIn(userInput, [fieldNameForPlaceholder])
-      return [...conditionArray, { text: textSegment }, { variable: inputtedData }]
-    }
-  }
-
-  function injectMultiFieldsAsObject(userInput, conditionText, userErrors, config) {
-    const variableString = (variable, error) => (error ? `[${error}]` : variable)
-
-    const strings = config.fields.map((fieldName) => {
-      return variableString(getIn(userInput, [fieldName]), getIn(userErrors, [fieldName]))
-    })
-
-    const fieldErrors = config.fields.map((fieldName) => getIn(userErrors, [fieldName])).filter((e) => e)
-
-    const string = interleave(strings, config.joining)
-
-    const variableKey = fieldErrors.length > 0 ? 'error' : 'variable'
-
-    const splitConditionText = conditionText.split(betweenBrackets).filter((text) => text)
-    return [{ text: splitConditionText[0] }, { [variableKey]: string }, { text: splitConditionText[1] }]
-  }
-
-  const betweenBrackets = /\[[^\]]*]/g
-
   const inputsFor = (fieldPositions, formInputs) => {
     const conditionAttributes = Object.keys(fieldPositions)
 
@@ -282,11 +126,9 @@ export default function createConditionsService() {
     }, {})
   }
 
-  const orderForView = (requiredOrder) => (a, b) => requiredOrder.indexOf(a) - requiredOrder.indexOf(b)
-
   return {
     // condition routes / review routes / pdf generation
-    populateLicenceWithConditions,
+    populateLicenceWithConditions: builder.populateLicenceWithConditions,
 
     // form generation
     getFullTextForApprovedConditions,
@@ -297,10 +139,6 @@ export default function createConditionsService() {
     formatConditionInputs,
     getStandardConditions,
     getAdditionalConditions,
-
-    // public only for testing
-    populateAdditionalConditionsAsObject,
-    populateLicenceWithApprovedConditions,
   }
 }
